update nbsphinx functions based on review feedback

+++ src/pydata_sphinx_theme/nbsphinx.py
@@ -7,8 +7,10 @@ from sphinx.application import Sphinx
 from . import utils
 
 
-# This function will allow us to catch during development or CI if nbsphinx ever
-# changes the name of the CSS file where it outputs its notebook styles.
+nbsphinx_css_filename = "nbsphinx-code-cells.css"
+replacement_css_filename = "nbsphinx-pydata-theme.css"
+
+
 def delete_nbsphinx_css(app: Sphinx) -> None:
     """For projects using nbsphinx, delete nbsphinx's CSS.
     We replace it later with our own.
@@ -16,9 +18,22 @@ def delete_nbsphinx_css(app: Sphinx) -> None:
     if "nbsphinx" not in app.config.extensions:
         return
 
-    nbsphinx_css = Path(app.builder.outdir) / "_static" / "nbsphinx-code-cells.css"
-    assert nbsphinx_css.exists()
-    nbsphinx_css.unlink()
+    nbsphinx_css = Path(app.builder.outdir) / "_static" / nbsphinx_css_filename
+    if nbsphinx_css.exists():
+        nbsphinx_css.unlink()
+    else:
+        # Here is the main purpose of this function. The main purpose is not so
+        # much to delete the nbsphinx CSS file (despite this function's name) as
+        # to help us catch a breaking change on nbsphinx's side if they ever
+        # rename or move its CSS file. This is important because the other
+        # function in this file, which gets executed later in the build process,
+        # also depends on nbsphinx's CSS being at a particular path. If not it's
+        # not there, then without this warning the CSS replacement mechanism
+        # would fail silently, and projects using this theme would load
+        # nbsphinx's notebook styles instead of this theme's.
+        utils.maybe_warn(
+            app, f"nbpshinx CSS not found in expected place: {nbsphinx_css}"
+        )
 
 
 def point_nbsphinx_pages_to_our_css(
@@ -36,7 +51,7 @@ def point_nbsphinx_pages_to_our_css(
     if "css_files" not in context:
         return
 
-    nbsphinx_css = "_static/nbsphinx-code-cells.css"
+    nbsphinx_css = "_static/" + nbsphinx_css_filename
     found = utils._delete_from_css_files(context["css_files"], nbsphinx_css)
     if found:
-        app.add_css_file("styles/nbsphinx-pydata-theme.css")
+        app.add_css_file("styles/" + replacement_css_filename)
